% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autokey.R
\name{autokey}
\alias{autokey}
\title{Autokey Cipher}
\usage{
autokey(message, key, encrypt = TRUE)
}
\arguments{
\item{message}{A character vector composed of a-z alphabetic characters}

\item{key}{An encryption key composed of a-z alphabetic characters}

\item{encrypt}{(Default: \code{TRUE}) TRUE will encrypt the message, while FALSE will decrypt the message.}
}
\value{
A character vector of either plaintext that has been encrypted or ciphertext that has been decrypted.
}
\description{
The Autokey Cipher is derived from the Vigenere Ciphere where the key and plaintext are bound to generate a new
encryption key as an input for the Vigenere method.
}
\examples{
autokey("VerySecretMessage", "Hack", encrypt = TRUE)
}
