% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columnar_transposition.R
\name{columnar_transposition}
\alias{columnar_transposition}
\title{Columnar Transposition Cipher}
\usage{
columnar_transposition(message, key, encrypt = TRUE)
}
\arguments{
\item{message}{A character vector}

\item{key}{A character vector used as the encryption key}

\item{encrypt}{(Default: \code{TRUE}) TRUE will encrypt the message, while FALSE will decrypt the message.}
}
\value{
A character vector of either plaintext that has been encrypted or ciphertext that has been decrypted.
}
\description{
Columnar Transposition Cipher
}
\examples{
columnar_transposition("Hidden message", "hack", encrypt = TRUE)
}
