% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{encode}
\alias{encode}
\title{Encode some text}
\usage{
encode(text)
}
\arguments{
\item{text}{Text to be encoded}
}
\value{
String of encoded text
}
\description{
Encode some text
}
\examples{
encode("this text will be encoded")
}
