% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{letter2num}
\alias{letter2num}
\title{Convert letter to number}
\usage{
letter2num(l)
}
\arguments{
\item{l}{A vector of lowercase letters.}
}
\value{
A vector of numbers from 1 to 27.
}
\description{
Converts a lowercase letter to its corresponding position in the alphabet
(e.g. a -> 1, z -> 26). Converts " " to 27.
}
\examples{
letter2num(c("h","e","l","l","o"))
}
