% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoder.R
\name{loglike}
\alias{loglike}
\title{Calculate the log-likelihood of this key based on probability matrix}
\usage{
loglike(textnum, y, M)
}
\arguments{
\item{textnum}{The encoded text, converted to a vector of chars}

\item{y}{The key that you're testing}

\item{M}{The probability matrix of likely two letter combinations}
}
\value{
The log-likelihood of this key
}
\description{
Calculate the log-likelihood of this key based on probability matrix
}
