% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{make_codable}
\alias{make_codable}
\title{Converts text with proper punctuation into text that is codable by converting
uppercase to lowercase and removing punctuation}
\usage{
make_codable(text)
}
\arguments{
\item{text}{Text that you're trying to make codable}
}
\value{
text that is now codable
}
\description{
Converts text with proper punctuation into text that is codable by converting
uppercase to lowercase and removing punctuation
}
\examples{
make_codable("Hello. My name is Oliver; what's your name?")
}
