% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode.R
\name{decode}
\alias{decode}
\title{Text Decoder}
\usage{
decode(code, iter = 7000, M = NULL, key = NULL)
}
\arguments{
\item{code}{An encoded .txt file. Make sure all letters within are lowercase.
Coded text should have the ` character rather than spaces.}

\item{iter}{Number of iterations to run the algorithm for.}

\item{M}{A 27x27 probability matrix of likely two-letter combinations.
Defaults to a probability matrix using 2-27 to represent letters and 1 to
represent the ` character.}
}
\value{
Decoded message in string form.
}
\description{
Decodes a text using the Metropolis algorithm.
}
\examples{
decode("~/Documents/encodedtext1.txt", iter = 10000)
}
