% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoder.R
\name{decoder}
\alias{decoder}
\title{Text Decoder}
\usage{
decoder(code, iter, M)
}
\arguments{
\item{code}{An encoded .txt file. Make sure all letters within are lowercase.
Decoded test will preserve periods (.).}

\item{iter}{Number of iterations to run the algorithm for.}

\item{M}{The probability matrix of likely two-letter combinations.}
}
\value{
Decoded message in string form.
}
\description{
Decodes a text using the Metropolis algorithm.
}
